Add header file for client display command
