+++ ut-robomaster/src/control/client-display/command_client_display.hpp
@@ -1,0 +1,73 @@
+#pragma once
+
+#include "tap/drivers.hpp"
+#include "tap/control/command.hpp"
+#include "tap/communication/serial/ref_serial_transmitter.hpp"
+#include "tap/communication/serial/ref_serial_data.hpp"
+#include "modm/processing/protothread.hpp"
+#include "modm/processing/resumable.hpp"
+
+#include "subsystems/odometry/odometry_subsystem.hpp"
+
+using namespace tap::control;
+using namespace tap::communication::serial;
+
+using subsystems::odometry::OdometrySubsystem;
+
+class BeybladeIndicator : protected modm::Resumable<2>
+{
+private:
+	const uint8_t* graphicName = (uint8_t*)"\x00\x00\x01";
+	RefSerialTransmitter &refSerialTransmitter;
+	RefSerialData::Tx::Graphic5Message msg;
+
+public:
+	BeybladeIndicator(RefSerialTransmitter refSerialTransmitter)
+	: refSerialTransmitter(refSerialTransmitter)
+	{}
+
+	modm::ResumableResult<bool> sendInitialGraphics();
+	modm::ResumableResult<bool> update();
+	
+	void initialize();
+};
+
+namespace commands
+{
+class CommandClientDisplay : Command, modm::pt::Protothread
+{
+private:
+	src::Drivers &drivers;
+	RefSerialTransmitter refSerialTransmitter;
+	BeybladeIndicator beybladeIndicator;
+	bool restarting = true;
+
+	void restartHud();
+public:
+	CommandClientDisplay(
+		src::Drivers drivers,
+		OdometrySubsystem *odometry) // temporarily using this to start this command
+		: Command(),
+		drivers(drivers),
+		refSerialTransmitter(&drivers),
+		beybladeIndicator(refSerialTransmitter)
+	{
+		addSubsystemRequirement(odometry);
+	}
+	
+	bool run();
+	
+	const char *getName() const override { return "client display"; }
+	void initialize() override
+	{
+		this->restarting = true;
+	};
+
+	void execute() override
+	{
+		run();
+	}
+	void end(bool) override {}
+	bool isFinished() const override { return false; }
+};
+}
